+++ src/components/MonthView.tsx
@@ -1,0 +1,98 @@
+import {
+  Stack,
+  Table,
+  TableBody,
+  TableCell,
+  TableContainer,
+  TableHead,
+  TableRow,
+  Typography,
+} from '@mui/material';
+
+import { WEEK_DAYS } from '../constants/events';
+import { Event } from '../types';
+import { CalendarEventItem } from './CalendarEventItem';
+import { formatDate, formatMonth, getEventsForDay, getWeeksAtMonth } from '../utils/dateUtils';
+
+interface MonthViewProps {
+  currentDate: Date;
+  filteredEvents: Event[];
+  notifiedEvents: string[];
+  holidays: Record<string, string>;
+}
+
+export const MonthView = ({
+  currentDate,
+  filteredEvents,
+  notifiedEvents,
+  holidays,
+}: MonthViewProps) => {
+  const weeks = getWeeksAtMonth(currentDate);
+
+  return (
+    <Stack data-testid="month-view" spacing={4} sx={{ width: '100%' }}>
+      <Typography variant="h5">{formatMonth(currentDate)}</Typography>
+      <TableContainer>
+        <Table sx={{ tableLayout: 'fixed', width: '100%' }}>
+          <TableHead>
+            <TableRow>
+              {WEEK_DAYS.map((day) => (
+                <TableCell key={day} sx={{ width: '14.28%', padding: 1, textAlign: 'center' }}>
+                  {day}
+                </TableCell>
+              ))}
+            </TableRow>
+          </TableHead>
+          <TableBody>
+            {weeks.map((week, weekIndex) => (
+              <TableRow key={weekIndex}>
+                {week.map((day, dayIndex) => {
+                  const dateString = day ? formatDate(currentDate, day) : '';
+                  const holiday = holidays[dateString];
+
+                  return (
+                    <TableCell
+                      key={dayIndex}
+                      sx={{
+                        height: '120px',
+                        verticalAlign: 'top',
+                        width: '14.28%',
+                        padding: 1,
+                        border: '1px solid #e0e0e0',
+                        overflow: 'hidden',
+                        position: 'relative',
+                      }}
+                    >
+                      {day && (
+                        <>
+                          <Typography variant="body2" fontWeight="bold">
+                            {day}
+                          </Typography>
+                          {holiday && (
+                            <Typography variant="body2" color="error">
+                              {holiday}
+                            </Typography>
+                          )}
+                          {getEventsForDay(filteredEvents, day).map((event) => {
+                            const isNotified = notifiedEvents.includes(event.id);
+                            return (
+                              <CalendarEventItem
+                                key={event.id}
+                                event={event}
+                                isNotified={isNotified}
+                              />
+                            );
+                          })}
+                        </>
+                      )}
+                    </TableCell>
+                  );
+                })}
+              </TableRow>
+            ))}
+          </TableBody>
+        </Table>
+      </TableContainer>
+    </Stack>
+  );
+};
